directions work
